requirements.txt
langgraph
langchain-openai
langchain-core
requests
python-dotenv
newsapi-python